refactor: break up App into main.rs, rm utils.rs
