--- src/utils.rs
@@ -1,3 +1,0 @@
-pub(crate) fn unescape(s: &str) -> Option<String> {
-    unescape::unescape(s)
-}
